+++ src/lib/constants.ts
@@ -21,4 +21,4 @@ export const limitWidth = cn(
   "mx-auto w-full max-w-[800px] lg:max-w-[1280px] px-4 md:px-7 lg:px-10"
 );
 
-export const termRevised = new Date(2024, 7 - 1, 1);
+export const termsRevised = new Date(2024, 7 - 1, 1);
